Taller-Procesos y Señales

+++ Talleres/Taller-Procesos y Señales/src/ejercicio1.c
@@ -17,10 +17,10 @@ int generate_random_number(){
 pid_t childs[9];
 
 void sigterm_handler(int signum){
-  sleep(1);
+  sleep(3);
+  pid_t pid = getpid();
   int randnum = generate_random_number();
   printf("El numero random del proceso %d: %d\n", getpid(), randnum);
-  pid_t pid = getpid();
   if(randnum == J){
     printf("Soy el proceso %d y me estoy por morir\n", pid); 
     exit(0);
@@ -30,8 +30,16 @@ void sigterm_handler(int signum){
 void sigchild_handler(int signum){
     pid_t pid;
     int status;
-
-    while ((pid = waitpid(-1, &status, WNOHANG)) > 0) {
+    // La función waitpid con la opción WNOHANG permite al proceso padre "consultar" la tabla de procesos 
+    // y ver si hay alguna entrada de un proceso hijo que ha terminado pero cuyo estado aún no ha sido recogido. 
+    // Si encuentra uno, lo recoge y retorna su PID y entro al while. Si no encuentra ninguno, simplemente retorna 0
+    // y no entro. Cada vez que paso por el ciclo verifico si hay un hijo que terminó. 
+    // El while es porque puede haber mas de un hijo que cambio de estado. Con un if 
+    // corro el riesgo de no poder agarrar la señal y que se me quede algun hijo zombie. 
+    // Ademas en UNIX si estoy atendiendo una señal del mismo tipo que me llega no se acumula sino que 
+    // continuo atendiendo la actual (que es la misma que me llego). 
+    while ((pid = waitpid(-1, &status, WNOHANG)) > 0) { 
+        sleep(1);
         for(int i = 0; i < N; i++) {
             if (childs[i] == pid) {
                 childs[i] = -1; 
@@ -73,14 +81,17 @@ int main(int argc, char const *argv[]){
         for(int j = 0; j < N; j++){
             sleep(1);
             pid = childs[j];
-            printf("Le mando una señal a mi hijo con pid: %d\n", pid);
-            kill(pid, SIGTERM); 
+            if(pid != -1){
+                printf("Le mando una señal a mi hijo con pid: %d\n", pid);
+                kill(pid, SIGTERM); 
+            }
         }   
     }
 
     for(int i = 0; i < N; i++){
         pid = childs[i];
         if (pid != -1){
+            sleep(3);
             printf("El proceso con id %d sobrevivio!\n", pid); 
             kill(pid, SIGKILL); 
         }
